harmless warning

+++ wrap/gl/gl_mesh_attributes_multi_viewer_bo_manager.h
@@ -296,7 +296,7 @@ namespace vcg
 			if (_glopts != NULL)
 			{
 				std::string tmp;
-				int size = _glopts->serialize(tmp);
+				size_t size = _glopts->serialize(tmp);
 				token[i] = str.substr(pos, size);
 				if (token[i].length() < size)
 					return false;
